More changes to graph relationships
