+++ src/lib/componentGraph.ts
@@ -5,16 +5,7 @@ import {FindCycles} from './DFSLib';
 import {AbstractGraph} from './abstractGraph';
 
 export class ComponentGraph extends AbstractGraph {
-    private _nodes: Map<string, Node> = new Map<string, Node>();
     private _groupMap: Map<ScalarNode, NodeGroup> = new Map<ScalarNode, NodeGroup>();
-
-    public get nodes(): IterableIterator<Node> {
-        return this._nodes.values();
-    }
-
-    public get nodeNames(): IterableIterator<string> {
-        return this._nodes.keys();
-    }
 
     public removeCycles(): void {
         let remover: FindCycles;
@@ -54,24 +45,17 @@ export class ComponentGraph extends AbstractGraph {
         return edges.values();
     }
 
-    public addEdge(src: Node, dst: Node): void {
-       (src as NodeImpl).addEdge(dst);
-    }
-    public getNode(name: string): Node {
-        return this._nodes.get(name);
-    }
-
     public combineNodes(node1: Node, node2: Node): Node {
-        assert.ok(this._nodes.has(node1.name), node1.name + ' doesn\'t exist');
-        assert.ok(this._nodes.has(node2.name), node2.name + ' doesn\'t exist');
+        assert.ok(this.nodesMap.has(node1.name), node1.name + ' doesn\'t exist');
+        assert.ok(this.nodesMap.has(node2.name), node2.name + ' doesn\'t exist');
         if (node1 === node2) {
             // already the same node, so nothing to do.
             return;
         }
         // combining nodes changes the name of the node, so we have to remove node1 and node2 from the
-        // _nodes map first, then put the maybe new node back.
-        this._nodes.delete(node1.name);
-        this._nodes.delete(node2.name);
+        // nodesMap map first, then put the maybe new node back.
+        this.nodesMap.delete(node1.name);
+        this.nodesMap.delete(node2.name);
         const ng: NodeGroup = (node1 as NodeImpl).combineWith(node2 as NodeImpl);
         if (node1 instanceof ScalarNode) {
             this._groupMap.set(node1, ng);
@@ -79,19 +63,8 @@ export class ComponentGraph extends AbstractGraph {
         if (node2 instanceof ScalarNode) {
             this._groupMap.set(node2, ng);
         }
-        this._nodes.set(ng.name, ng);
+        this.nodesMap.set(ng.name, ng);
         return ng;
-    }
-
-    public getOrAddNode(name: string, details: Map<string, object>): Node {
-        let n: Node = this._nodes.get(name);
-        if (n) {
-            return n;
-        }
-
-        n = new ScalarNode(name, details);
-        this._nodes.set(name, n);
-        return n;
     }
 
 }
